+++ src/main/java/com/example/happyprogramming/service/implement/MentorServiceImpl.java
@@ -24,7 +24,7 @@ public class MentorServiceImpl implements MentorService {
     }
 
     @Override
-    public CVEntity findMentorById(int id) {
+    public CVEntity findMentorById(long id) {
         return cvRepository.findByUser(userRepository.findById(id));
     }
 }
